+++ packages/js-sdk/src/api/group.ts
@@ -1,10 +1,3 @@
-import { TxClient } from '../clients/txClient';
-import { MsgCreateGroupSDKTypeEIP712 } from '../messages/greenfield/storage/MsgCreateGroup';
-import { MsgDeleteGroupSDKTypeEIP712 } from '../messages/greenfield/storage/MsgDeleteGroup';
-import { MsgLeaveGroupSDKTypeEIP712 } from '../messages/greenfield/storage/MsgLeaveGroup';
-import { MsgUpdateGroupExtraSDKTypeEIP712 } from '../messages/greenfield/storage/MsgUpdateGroupExtra';
-import { getMsgUpdateGroupMemberSDKTypeEIP712 } from '../messages/greenfield/storage/MsgUpdateGroupMember';
-import { GRNToString, newBucketGRN, newGroupGRN, newObjectGRN } from '../utils/grn';
 import {
   QueryGroupNFTResponse,
   QueryHeadGroupMemberResponse,
@@ -34,6 +27,13 @@ import {
   TxResponse,
 } from '..';
 import { RpcQueryClient } from '../clients/queryclient';
+import { TxClient } from '../clients/txClient';
+import { getMsgCreateGroupSDKTypeWithTagEIP712 } from '../messages/greenfield/storage/MsgCreateGroup';
+import { MsgDeleteGroupSDKTypeEIP712 } from '../messages/greenfield/storage/MsgDeleteGroup';
+import { MsgLeaveGroupSDKTypeEIP712 } from '../messages/greenfield/storage/MsgLeaveGroup';
+import { MsgUpdateGroupExtraSDKTypeEIP712 } from '../messages/greenfield/storage/MsgUpdateGroupExtra';
+import { getMsgUpdateGroupMemberSDKTypeEIP712 } from '../messages/greenfield/storage/MsgUpdateGroupMember';
+import { GRNToString, newBucketGRN, newGroupGRN, newObjectGRN } from '../utils/grn';
 import { Storage } from './storage';
 
 export interface IGroup {
@@ -111,6 +111,10 @@ export class Group implements IGroup {
   private queryClient: RpcQueryClient = container.resolve(RpcQueryClient);
 
   public async createGroup(msg: MsgCreateGroup) {
+    const isTagsEmpty = msg?.tags?.tags?.length === 0;
+
+    const MsgCreateGroupSDKTypeEIP712 = getMsgCreateGroupSDKTypeWithTagEIP712(isTagsEmpty);
+
     return await this.txClient.tx(
       MsgCreateGroupTypeUrl,
       msg.creator,