+++ packages/reed-solomon/types/web.adapter.d.ts
@@ -1,0 +1,5 @@
+declare module '@bnb-chain/reed-solomon/web.adapter' {
+  export class WebAdapterReedSolomon {
+    encodeInWorker(workerFn: () => void, data: Uint8Array): Promise<string[]>;
+  }
+}